requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.1.0
langgraph==0.0.26
openai==1.3.7
langchain-openai==0.0.5
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
fastmcp==0.9.0
mcp==1.0.0
typing-extensions==4.8.0